Move LazyImage into 'utilities' folder.
